move around creatables
